# Optimize Hex.of_string/Hex.to_string