--- a/lib/hex.ml
+++ b/lib/hex.ml
@@ -52,10 +52,11 @@ let of_string_fast s =
   let len = String.length s in
   let buf = Bytes.create (len * 2) in
   for i = 0 to len - 1 do
+    let c = String.unsafe_get s i |> Char.code in
     Bytes.unsafe_set buf (i * 2)
-      (String.unsafe_get hexa1 (Char.code (String.unsafe_get s i)));
+      (String.unsafe_get hexa1 c);
     Bytes.unsafe_set buf (succ (i * 2))
-      (String.unsafe_get hexa2 (Char.code (String.unsafe_get s i)));
+      (String.unsafe_get hexa2 c)
   done;
   `Hex (Bytes.to_string buf)