Although the compiler should (eventually) be able to do these optimisations itself it is worthwhile to optimise these commonly used functions by hand for now, and eliminating the extra copies (reminiscent of the safe string changs) reduces the amount of memory allocated and the work the GC has to do.

Before:
```
╭──────────────────────────────────────────┬───────────────────────────┬───────────────────────────┬───────────────────────────╮
│name                                      │  major-allocated          │  minor-allocated          │  monotonic-clock          │
├──────────────────────────────────────────┼───────────────────────────┼───────────────────────────┼───────────────────────────┤
│  Encodings.validate/Hex decode:10        │             0.0000 mjw/run│            15.0491 mnw/run│            125.3125 ns/run│
│  Encodings.validate/Hex decode:1000      │             0.0000 mjw/run│           139.2018 mnw/run│           6543.7140 ns/run│
│  Encodings.validate/Hex decode:10000     │          1254.0000 mjw/run│            11.5263 mnw/run│          64148.3848 ns/run│
│  Encodings.validate/Hex encode:10        │             0.0000 mjw/run│            11.0153 mnw/run│             65.2861 ns/run│
│  Encodings.validate/Hex encode:1000      │             0.2297 mjw/run│           507.1115 mnw/run│           3426.7689 ns/run│
│  Encodings.validate/Hex encode:10000     │          5004.0000 mjw/run│             3.4478 mnw/run│          44723.6380 ns/run│
╰──────────────────────────────────────────┴───────────────────────────┴───────────────────────────┴───────────────────────────╯
```

After:
```
╭──────────────────────────────────────────┬───────────────────────────┬───────────────────────────┬───────────────────────────╮
│name                                      │  major-allocated          │  minor-allocated          │  monotonic-clock          │
├──────────────────────────────────────────┼───────────────────────────┼───────────────────────────┼───────────────────────────┤
│  Encodings.validate/Hex decode:10        │             0.0000 mjw/run│            12.0397 mnw/run│             62.8569 ns/run│
│  Encodings.validate/Hex decode:1000      │             0.0000 mjw/run│            74.1665 mnw/run│           4242.3554 ns/run│
│  Encodings.validate/Hex decode:10000     │           627.0000 mjw/run│            10.4323 mnw/run│          38059.3616 ns/run│
│  Encodings.validate/Hex encode:10        │             0.0000 mjw/run│             7.0149 mnw/run│             41.3435 ns/run│
│  Encodings.validate/Hex encode:1000      │             0.0000 mjw/run│           255.1058 mnw/run│           2686.6025 ns/run│
│  Encodings.validate/Hex encode:10000     │          2502.0000 mjw/run│             3.3778 mnw/run│          27802.6243 ns/run│
╰──────────────────────────────────────────┴───────────────────────────┴───────────────────────────┴───────────────────────────╯
```

(Measured on Intel(R) Xeon(R) Silver 4108 CPU @ 1.80GHz with OCaml 4.13.1)

I was a bit surprised to find out that nested functions prevent inlining (at least with a default compiler that doesn't have flambda enabled), and hoisting the function out manually made inlining possible (I think even without flambda the compiler is meant to do the hoisting optimization, but perhaps too late after inlining?)